feat: experimental client server, finished version incoming
I created a small server to attack that can send his public key and
decrypt messages. When it decrypt messages it only send if things went
wrong or not, it also check PKCS1v15.

The client is here to send get the key and send back encrypted messages
right now, later it will do the analysis part too.

The chanegs on rsaTools were anly completed some type so it will be
easier to work with it in other files.