feat: deletion of server and real implementation to finding valid padding
Since the group have decided to do everything in one file I've deleted
the server file and created changes the client one we will be rename
later. In it, based on Everlasting ROBOTS paper I started implementing
the code.

There is for the moment the oracle. It is here to answer if the padding
is good or not. And the valid padding finder which find out the good
padding for the code.